# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cobranca-API — Offline (engine pyCobranca)
  description: |
    Superfície **offline** da plataforma **Cobranca-API** (serviço único **100% Python**):
    geração de boletos bancários brasileiros, arquivos CNAB e parsing de extratos OFX —
    servida **nativamente** pela engine [pyCobrança](https://github.com/Maxwbh/pyCobranca)
    dentro do próprio processo FastAPI (sem sidecar, sem proxy).

    > **Nota de versão (v2.0.0):** os contratos desta superfície (paths, campos,
    > headers `X-*`) permanecem compatíveis com a v1.x. A única mudança de escopo é
    > o formato de saída: a engine emite **PDF** — `jpg`/`png`/`tif` respondem
    > **400**.

    ## Offline (`/api/*`) × Gateway (raiz) — quando usar cada um

    | | **Offline (esta doc)** | **Gateway REST** ([Swagger `/docs`](/docs)) |
    |---|---|---|
    | O que faz | GERA o documento (boleto/CNAB) localmente | REGISTRA a cobrança na API online do banco (C6, Sicoob, Inter, Itaú) |
    | Boleto | Não registrado (o banco só conhece no pagamento) | **Registrado** (CIP, em tempo real) + Pix, Bolepix, webhooks |
    | Autenticação | **Nenhuma** (por design — não toca em banco nem segredo) | Token `bapi_` (`POST /credenciais`) |
    | Conciliação | Retorno CNAB / OFX (arquivos) | Webhooks + consultas online |

    > **Autenticação:** esta superfície não tem `securitySchemes` de propósito.
    > O mecanismo `bapi_` pertence ao gateway e está documentado no
    > [Swagger do gateway](/docs).

...

...

...

...

...
servers:
...
paths:
...
  /api/boleto/multi:
    post:
      tags:
        - Boleto
      summary: Gerar múltiplos boletos · 📄 resposta binária
      description: |
        Gera múltiplos boletos em um único arquivo.

        **Lote resiliente:** a falha de um item **não** cancela o lote. O PDF sai
        com os itens válidos e o resultado por item vem nos headers
        `X-Boletos-Info` (ok), `X-Boletos-Errors`, `X-Boletos-Failed` e
        `X-Batch-Status` (`completed` | `partially_completed`). Com
        `include_data=true`, o JSON traz `{status, total, completed, failed,
        boletos, erros}`. **400** só quando nenhum item é válido.

        ⚠️ **Limite**: `LOTE_MAX_ITENS` (default **200**) — acima disso responde
        **413**; para lotes maiores use processamento assíncrono
        (ver `docs/development/plano-jobs-lote.md`).
      operationId: generateMultipleBoletos
      parameters:
        # Query, não campo do formulário: o corpo multipart carrega SÓ o
        # arquivo, como em /api/remessa e /api/retorno. Esta página já disse o
        # contrário, e o efeito era silencioso — `include_data` mandado no
        # formulário era ignorado e a resposta vinha binária.
        - name: type
          in: query
          required: false
          description: Formato de saída (só `pdf`; qualquer outro responde 400).
          schema:
            $ref: '#/components/schemas/OutputType'
        - name: include_data
          in: query
          required: false
          description: |
            Se `true`, retorna JSON com os metadados de cada boleto + o PDF
            combinado em base64. Caso contrário, retorna o binário + os
            headers `X-Boletos-*`.
          schema:
            type: string
            enum: ['true', 'false']
            default: 'false'
        - name: template
          in: query
          required: false
          description: |
            Modelo visual. Além de `moderno`/`classico`, esta rota aceita
            `carne` (3 vias por A4) — que `GET /api/boleto` recusa.
          schema:
            type: string
            enum: [moderno, classico, carne]
            default: moderno
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - data
              properties:
                data:
                  type: string
                  format: binary
                  description: |
                    Arquivo JSON com um **array** de boletos, cada um com o
                    campo `bank`. Outra forma (objeto, texto, número) responde
                    **400** dizendo o que chegou.
      responses:
        '200':
          description: |
            Sem include_data: arquivo binário + headers `X-Boletos-Info` e `X-Boletos-Count`.
            Com include_data=true: JSON contendo array `boletos` + `content_base64`.
          headers:
            X-Boletos-Count:
              schema:
                type: integer
              description: |
                Boletos **gerados** — não o tamanho do lote enviado. Num lote
                de 3 com 1 falha, vale 2 (e `X-Boletos-Failed` vale 1).
              example: 2
            X-Boletos-Info:
              schema:
                type: string
                format: json
              description: |
                Array JSON com o registro de cada boleto gerado: `item_id`,
                `indice`, `bank`, `status`, `nosso_numero`,
                `nosso_numero_formatado`, `nosso_numero_dv`, `codigo_barras` e
                `linha_digitavel`.
              example: '[{"item_id":"0","indice":0,"bank":"banco_brasil","status":"completed","nosso_numero":"123","nosso_numero_formatado":"12345670000000123","nosso_numero_dv":"","codigo_barras":"...","linha_digitavel":"..."}]'
            X-Boletos-Failed:
              schema:
                type: integer
              description: Quantidade de itens que falharam na validação.
              example: 1
            X-Boletos-Errors:
              schema:
                type: string
                format: json
              description: |
                Array JSON com os itens que falharam (`item_id`, `indice`,
                `bank`, `status`, `errors`). `[]` quando não há falhas.
              example: '[{"item_id":"2","indice":2,"bank":"itau","status":"failed","errors":["conta_corrente invalido"]}]'
            X-Batch-Status:
              schema:
                type: string
                enum: [completed, partially_completed]
              description: Resultado global do lote.
              example: partially_completed
            X-Boletos-Info-Truncado:
              schema:
                type: string
                enum: ['true']
              description: |
                Presente **só** quando `X-Boletos-Info` não coube no header e
                virou `[]`. O PDF continua sendo entregue e os contadores
                (`X-Boletos-Count`, `X-Boletos-Failed`) seguem válidos —
                para o detalhe por item, use `include_data=true`.

                Por que existe: o metadado cresce ~300 B por boleto (100 itens
                ≈ 30 KB, 200 ≈ 60 KB). Acima de ~64 KB o `http.client` do Python
                recusa a resposta **inteira** e o cliente perde até o PDF.
                Com `LOTE_MAX_ITENS` no default (200) isso **nunca** dispara;
                ajuste `HEADER_JSON_MAX_BYTES` (ex.: `6144`) se estiver atrás de
                nginx (8 KB) ou ALB (16 KB).
            X-Boletos-Errors-Truncado:
              schema:
                type: string
                enum: ['true']
              description: |
                O mesmo que `X-Boletos-Info-Truncado`, para `X-Boletos-Errors`.
            X-Boletos-Detalhe:
              schema:
                type: string
              description: Onde buscar o detalhe quando houve truncagem.
              example: use include_data=true (metadados no corpo)
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                $ref: '#/components/schemas/MultiBoletoResponse'
              examples:
                with_include_data:
                  summary: Response com include_data=true (lote de 3, 1 falha)
                  value:
                    total: 3
                    completed: 2
                    failed: 1
                    status: partially_completed
                    boletos:
                      - item_id: '0'
                        indice: 0
                        bank: banco_brasil
                        status: completed
                        nosso_numero: '123'
                        nosso_numero_formatado: '12345670000000123'
                        nosso_numero_dv: ''
                        codigo_barras: '00198204100000150000000001234567000000012318'
                        linha_digitavel: '00190.00009 01234.567004 00000.123182 8 20410000015000'
                      - item_id: '1'
                        indice: 1
                        bank: sicoob
                        status: completed
                        nosso_numero: '7890'
                        nosso_numero_formatado: '00078900'
                        nosso_numero_dv: ''
                        codigo_barras: '75691204100000250001432701022938500078900001'
                        linha_digitavel: '75691.43279 01022.938508 00789.000015 1 20410000025000'
                    erros:
                      - item_id: '2'
                        indice: 2
                        bank: itau
                        status: failed
                        errors:
                          - "conta deve ter no máximo 5 dígitos"
                    content_base64: 'JVBERi0xLjQK...'
                    content_type: application/pdf
                    filename: boletos-multi.pdf
        '400':
          description: |
            Nenhum item do lote é válido, `data` não é um array de objetos, ou
            `type`/`template` fora dos valores aceitos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: |
            Lote acima do limite `LOTE_MAX_ITENS` (default **200**), ou arquivo
            acima de `UPLOAD_MAX_BYTES` (default **10 MB**). O teto de bytes é
            conferido ANTES de carregar o conteúdo.
            Para volumes maiores, use o processamento assíncrono
            (`POST /jobs/boletos`, no gateway).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Lote acima do limite de 200 itens
                  recebidos:
                    type: integer
                    example: 350
                  hint:
                    type: string
        '422':
          $ref: '#/components/responses/LoteDuplicado'
...
  /api/render/carne:
    post:
      tags:
        - Render
      summary: Renderizar carnê (N boletos) em PDF
      description: |
        Gera o carnê a partir de uma lista de boletos, no layout `carne`
        (3 vias por A4) da engine pyCobrança.

        ⚠️ **Timeout**: carnês com muitas parcelas levam vários segundos —
        timeout do cliente ≥ 60s.
      operationId: renderCarne
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [boletos]
              properties:
                bank:
                  type: string
                  description: Banco comum a todos (se cada boleto não trouxer o seu)
                  example: itau
                boletos:
                  type: array
                  items: { type: object }
                  description: |
                    Lista de boletos (cada um com seus dados). Outra forma —
                    ou item que não seja objeto — responde **400** dizendo o
                    que chegou.
                template:
                  type: string
                  enum: [carne]
                  default: carne
      responses:
        '200':
          description: Carnê renderizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  pdf_base64: { type: string, description: PDF do carnê em base64 }
        '400':
          description: Falha ao gerar carnê
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: |
            Lote acima do limite `LOTE_MAX_ITENS` (default **200**). O teto é o
            mesmo de `/api/boleto/multi` — as duas rotas renderizam pelo mesmo
            caminho, de forma síncrona.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string, example: Lote acima do limite de 200 itens }
                  recebidos: { type: integer, example: 350 }
                  hint: { type: string }
        '422':
          $ref: '#/components/responses/LoteDuplicado'
...
components:
  responses:
    LoteDuplicado:
      description: |
        Dois itens com o mesmo identificador — o `item_id` sai de
        `external_id`, `seu_numero` ou `numero_documento`. É o mesmo título
        emitido duas vezes: o segundo sobrescreve o primeiro e some do lote
        sem aviso, então o lote é recusado inteiro.
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: Itens com identificador duplicado no lote }
              hint: { type: string }
              duplicados:
                type: array
                items:
                  type: object
                  properties:
                    item_id: { type: string, example: 'NF-1042' }
                    indices:
                      type: array
                      items: { type: integer }
                      example: [0, 3]
  schemas:
    BankCode:
      type: string
      enum:
        - banco_brasil
        - itau
        - bradesco
        - caixa
        - santander
        - sicoob
        - sicredi
        - banrisul
        - banestes
        - banco_nordeste
        - banco_brasilia
        - banco_c6
        - unicred
        - credisis
        - safra
        - citibank
        - hsbc
        - ailos
      description: Código do banco suportado (18 bancos brasileiros)
    OutputType:
      type: string
      enum:
        - pdf
      default: pdf
      description: |
        Formato de saída do boleto. A engine pyCobrança emite **PDF**;
        `jpg`/`png`/`tif` (v1.x) respondem **400**.
...
    MultiBoletoResponse:
      type: object
      description: |
        Response retornada quando POST /api/boleto/multi é chamado com `include_data=true`.
      properties:
        total:
          type: integer
          description: Itens **recebidos** no lote (válidos + falhos).
          example: 3
        completed:
          type: integer
          description: Itens que viraram boleto.
          example: 2
        failed:
          type: integer
          description: Itens recusados na validação.
          example: 1
        status:
          type: string
          enum: [completed, partially_completed]
          description: |
            `partially_completed` quando ao menos um item falhou. Lote sem
            nenhum item válido não chega aqui — responde 400.
          example: partially_completed
        boletos:
          type: array
          description: Um registro por item gerado.
          items:
            $ref: '#/components/schemas/BoletoMetadata'
        erros:
          type: array
          description: |
            Um registro por item recusado — o lote não é cancelado por causa
            deles. Mesmo conteúdo do header `X-Boletos-Errors`.
          items:
            $ref: '#/components/schemas/BoletoFalho'
        content_base64:
          type: string
          format: base64
          description: Arquivo combinado (multi-página) em base64
        content_type:
          type: string
          example: application/pdf
        filename:
          type: string
          example: boletos-multi.pdf
    BoletoMetadata:
      type: object
      description: Metadados de um boleto no array `boletos` ou no header `X-Boletos-Info`
      properties:
        item_id:
          type: string
          description: |
            Identificador do item — sai de `external_id`, `seu_numero` ou
            `numero_documento`; na falta dos três, do índice.
          example: '0'
        indice:
          type: integer
          description: Posição no lote enviado.
          example: 0
        status:
          type: string
          enum: [completed]
        bank:
          $ref: '#/components/schemas/BankCode'
        nosso_numero:
          type: string
          example: '123'
        nosso_numero_formatado:
          type: string
          example: '12345670000000123'
        nosso_numero_dv:
          type: string
          example: ''
        codigo_barras:
          type: string
        linha_digitavel:
          type: string
    BoletoFalho:
      type: object
      description: |
        Item recusado no lote, no array `erros` ou no header `X-Boletos-Errors`.
      properties:
        item_id:
          type: string
          example: '2'
        indice:
          type: integer
          example: 2
        bank:
          type: string
          example: itau
        status:
          type: string
          enum: [failed]
        errors:
          type: array
          items:
            type: string
          example: ["conta deve ter no máximo 5 dígitos"]
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Mensagem de erro
          example: Dados inválidos
        validation_errors:
          type: array
          description: Lista plana — um item por campo violado (contrato PyCobrança ≥ 1.0.1).
          items:
            type: string
        details:
          type: array
          description: |
            Só em `/api/retorno`, que usa este nome no lugar de
            `validation_errors` — um item por problema do arquivo.
          items:
            type: string
        hint:
          type: string
          description: O que fazer, quando a rota tem o que sugerir.
          example: Verifique se todos os campos obrigatórios estão preenchidos